src/function-app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
Scrapy==2.13.3
markdownify==1.2.0
azure-storage-blob==12.26.0
azure-identity==1.18.0